requirements.txt
pandas>=2.2
numpy>=1.26
requests>=2.31
pydantic>=2.7
python-dotenv>=1.0
yfinance>=0.2
duckdb>=1.0
beautifulsoup4>=4.12
